Update to use getAnswerRows
